requirement.txt
fastapi
uvicorn[standard]
pydantic>=2
sqlalchemy>=2
python-dotenv
google-generativeai>=0.5     
openai-whisper
sounddevice                 
coqui-tts
python-multipart             # 음성 업로드용